requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
feedparser==6.0.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.68.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==22.0
requests==2.32.3
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.6
telegram==0.0.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
